requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
passlib==1.7.4
pluggy==1.4.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==2.6.2
pydantic_core==2.16.3
pytest==8.0.1
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
schedule==1.2.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
tomli==2.0.1
typing_extensions==4.9.0
uvicorn==0.27.1
